requirements.txt
streamlit
torch
torchvision
pillow
numpy
pennylane
medmnist